# server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Foodsharing EinAb API
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/public/bezirke/{bezirkSlug}/slots/available:
    get:
      tags: [public]
      operationId: getAvailableSlots
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/EinAbCategory'
        - name: visitFairteiler
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Available bookable slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableSlotListResponse'
  /api/public/bezirke/{bezirkSlug}/slots/{slotId}/book:
    post:
      tags: [public]
      operationId: bookSlot
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - $ref: '#/components/parameters/SlotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookSlotRequest'
      responses:
        '200':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/public/bookings/confirm:
    post:
      tags: [public]
      operationId: confirmBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyTokenRequest'
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/public/bezirke/{bezirkSlug}/teachers/signup:
    post:
      tags: [public]
      operationId: signupTeacher
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherSignupRequest'
      responses:
        '201':
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/public/bezirke/{bezirkSlug}/notifications/subscribe:
    post:
      tags: [public]
      operationId: subscribeNotifications
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSubscriptionRequest'
      responses:
        '200':
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/public/notifications/unsubscribe:
    post:
      tags: [public]
      operationId: unsubscribeNotifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationUnsubscribeRequest'
      responses:
        '200':
          description: Subscription disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    BezirkSlug:
      name: bezirkSlug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9-]+$'
    SlotId:
      name: slotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [message, status, code]
      properties:
        status:
          type: integer
        message:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'
        details:
          type: array
          items:
            type: string
    ErrorCode:
      type: string
      enum: [AUTHENTICATION_REQUIRED, ACCESS_DENIED, VALIDATION_FAILED, RESOURCE_NOT_FOUND, UNEXPECTED_ERROR, BEZIRK_NOT_FOUND, USER_BEZIRK_MISMATCH, NOTIFICATION_SUBSCRIPTION_BEZIRK_MISMATCH, INVALID_LOGIN_TOKEN, LOGIN_TOKEN_EXPIRED_OR_USED, ACCOUNT_NOT_FOUND, INVALID_UNSUBSCRIBE_TOKEN, TEACHER_ACCOUNT_REQUIRED, TEACHER_FOODSHARING_ID_ALREADY_EXISTS, TEACHER_NOT_FOUND, TEACHER_INACTIVE, BOOKING_USER_DISABLED, BOOKING_USER_NOT_FOUND, SLOT_NOT_FOUND, BOOKING_NOT_FOUND, SLOT_NOT_AVAILABLE, USER_ALREADY_BOOKED_WITH_TEACHER, USER_ALREADY_BOOKED_IN_CATEGORY, USER_BOOKING_LIMIT_REACHED, MINIMUM_PICKUP_COUNT_NOT_REACHED, ONLY_OWN_BOOKINGS_CANCELLABLE, ONLY_BOOKED_APPOINTMENTS_CANCELLABLE, EINAB_NOT_FOUND, ONLY_OWN_EINABS_MANAGEABLE, SLOT_COUNT_OUT_OF_RANGE, SLOT_COUNT_REDUCTION_NOT_SUPPORTED, BOOKED_SLOTS_PREVENT_DELETE, PUBLIC_LOCATION_REQUIRED, DEFAULT_ADMIN_CANNOT_BE_CHANGED, INVALID_BOOKING_CONFIRMATION_TOKEN, BOOKING_CONFIRMATION_EXPIRED]
    BezirkResponse:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
...
    Language:
      type: string
      enum: [de, en, gws]
    EinAbCategory:
      type: string
      enum: [SUPERMARKET, TAKEOUT, MARKET, BAKERY, RESTAURANT, FAIRTEILER_CLEANING]
    SlotStatus:
      type: string
      enum: [AVAILABLE, PENDING_CONFIRMATION, BOOKED, DONE, CANCELLED]
...
    BookSlotRequest:
      type: object
      required: [foodsharingId, language]
      properties:
        foodsharingId:
          type: string
          pattern: '^\d+$'
        language:
          $ref: '#/components/schemas/Language'
    TeacherSignupRequest:
      type: object
      required: [foodsharingId, language]
      properties:
        foodsharingId:
          type: string
          pattern: '^\d+$'
        icalLink:
          type: string
          format: uri
        language:
          $ref: '#/components/schemas/Language'
    NotificationSubscriptionRequest:
      type: object
      required: [email, language]
      properties:
        email:
          type: string
          format: email
        language:
          $ref: '#/components/schemas/Language'
    NotificationUnsubscribeRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
    NotificationSubscriptionResponse:
      type: object
      required: [email, active, bezirk]
      properties:
        email:
          type: string
          format: email
        active:
          type: boolean
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
...
    VerifyTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
...
    TeacherResponse:
      type: object
      required: [id, foodsharingId, name, active, language, createdAt, updatedAt, canGiveEinAbs, canManageUsers, canUseAutomations, canSeeUserPickupCountGrouping, canUseAutomationSlotApproval, canUseAutomationRequestApproval, canUseAutomationOpenSlotAdvertising, canSeeAllAutomationDecisions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        foodsharingId:
          type: string
        name:
          type: string
        phoneNumber:
          type: string
        icalLink:
          type: string
        active:
          type: boolean
        canGiveEinAbs:
          type: boolean
        canManageUsers:
          type: boolean
        canUseAutomations:
          type: boolean
        canSeeUserPickupCountGrouping:
          type: boolean
        canUseAutomationSlotApproval:
          type: boolean
        canUseAutomationRequestApproval:
          type: boolean
        canUseAutomationOpenSlotAdvertising:
          type: boolean
        canSeeAllAutomationDecisions:
          type: boolean
        wantsToBeTeacher:
          type: boolean
        language:
          $ref: '#/components/schemas/Language'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
...
    BookingUserResponse:
      type: object
      required: [id, name, foodsharingId, language, active, createdAt, updatedAt, canGiveEinAbs, canManageUsers, canUseAutomations, canSeeUserPickupCountGrouping, canUseAutomationSlotApproval, canUseAutomationRequestApproval, canUseAutomationOpenSlotAdvertising, canSeeAllAutomationDecisions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        foodsharingId:
          type: string
        phoneNumber:
          type: string
        language:
          $ref: '#/components/schemas/Language'
        active:
          type: boolean
        canGiveEinAbs:
          type: boolean
        canManageUsers:
          type: boolean
        canUseAutomations:
          type: boolean
        canSeeUserPickupCountGrouping:
          type: boolean
        canUseAutomationSlotApproval:
          type: boolean
        canUseAutomationRequestApproval:
          type: boolean
        canUseAutomationOpenSlotAdvertising:
          type: boolean
        canSeeAllAutomationDecisions:
          type: boolean
        wantsToBeTeacher:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
...
    AvailableSlotResponse:
      type: object
      required: [slotId, bezirk, category, startDateTime, teacherName, visitFairteiler, status]
      properties:
        slotId:
          type: string
          format: uuid
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
        category:
          $ref: '#/components/schemas/EinAbCategory'
        startDateTime:
          type: string
          format: date-time
        location:
          type: string
        publicLocation:
          type: string
        teacherName:
          type: string
        teacherId:
          type: string
          format: uuid
        minimumPickupCount:
          type: integer
        visitFairteiler:
          type: boolean
        status:
          $ref: '#/components/schemas/SlotStatus'
    AvailableSlotListResponse:
      type: object
      required: [slots, page, size, totalElements, totalPages]
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/AvailableSlotResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    BookingDetailResponse:
      type: object
      required: [slotId, einAbId, bezirk, category, startDateTime, teacherName, teacherPhoneNumber, status, bookingUser]
      properties:
        slotId:
          type: string
          format: uuid
        einAbId:
          type: string
          format: uuid
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
        category:
          $ref: '#/components/schemas/EinAbCategory'
        startDateTime:
          type: string
          format: date-time
        location:
          type: string
        publicLocation:
          type: string
        whatToBring:
          type: string
        hint:
          type: string
        teacherName:
          type: string
        teacherPhoneNumber:
          type: string
        visitFairteiler:
          type: boolean
        status:
          $ref: '#/components/schemas/SlotStatus'
        bookingUser:
          $ref: '#/components/schemas/BookingUserResponse'
...